+++ lib/markers/const/default-marker-type.const.dart
@@ -1,7 +1,7 @@
 import '../models/marker-type.model.dart';
 
 // In case no marker types have been defined there's a default type.
-final defaultMarkerType = MarkerTypeM(
+const DEFAULT_MARKER_TYPE = MarkerTypeM(
   id: 'reminder',
   name: 'Reminder',
 );